fix: revision handling for parallel upload to the same node
When multiple uploads on the same node are in post-processing at the
same time the upload that enters post-processing 2nd will created a
new file revision.
However if the upload finished post-processing before the 1st upload
the blob of the first upload ended up over-writing the original node.
We now make sure that the 1st upload's session is set to write the blob
to the revision node.

Partial fix for: https://github.com/opencloud-eu/opencloud/issues/1763