fix(auth/scope): Fix 'publicshare' auth scope validation
To verify that a resource is within the scope of a public link token,
both the resourceId from the scope and the accessed reference need to be
Stat'ed in order to check:
a) whether they're within the same space (by checking the space ids)
b) whether the accessed ref is actually a child of the shared resource
(by checking the paths)

Additionally the cache key for the 'scopeExpansionCache' is fixed to
to take the resource id of the target resource into account.
scope# This is the commit message #2:

With the above fixes the `checkRelativeRefernce()` method is no longer
needed and was removed. It contained another bug that would allow access
to resources out of the share scope.

Fixes: CVE-2026-23989